🛠️	Refactor : sample에 들어가 있는 Redis와 MongoDB JPA로부터 해방
더이상 Jpa가 Bean으로 등록하지 못하도록 리펙토링
