+++ src/main/java/darkoverload/itzip/sample/Repository/SampleRedisRepository.java
@@ -1,6 +1,6 @@
 package darkoverload.itzip.sample.Repository;
 
-import darkoverload.itzip.global.config.mongo.ExcludeFromJpaRepositories;
+import darkoverload.itzip.global.config.querydsl.ExcludeFromJpaRepositories;
 import darkoverload.itzip.sample.domain.SampleRedis;
 import org.springframework.data.repository.CrudRepository;
